// src/app/api/models/user.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Absence } from '../models/absence';
import { Company } from '../models/company';
import { EmploymentContract } from '../models/employment-contract';
import { Role } from '../models/role';
import { TimeEntry } from '../models/time-entry';
export interface User {
  absences?: Array<Absence> | null;
  company?: Company;
  companyId?: string;
  email?: string | null;
  employmentContract?: EmploymentContract;
  firstName?: string | null;
  id?: string;
  lastName?: string | null;
  passwordHash?: string | null;
  role?: Role;
  timeEntries?: Array<TimeEntry> | null;
}